+++ tests/test_websocket_decoder.c
@@ -178,7 +178,7 @@ DECODER_TEST_CASE(websocket_decoder_rsv) {
         };
 
         /* Set the appropriate RSV */
-        input[0] |= (1 << (6 - rsv));
+        input[0] |= (uint8_t)(1 << (6 - rsv));
 
         struct aws_websocket_frame expected_frame = {
             .fin = true,
